+++ common/src/main/java/de/maxhenkel/voicechat/plugins/impl/audiochannel/EntityAudioChannelImpl.java
@@ -61,7 +61,7 @@ public class EntityAudioChannelImpl extends AudioChannelImpl implements EntityAu
 
     private void broadcast(PlayerSoundPacket packet) {
         if (entity.getEntity() instanceof net.minecraft.world.entity.Entity entity) {
-            server.broadcast(ServerWorldUtils.getPlayersInRange((ServerLevel) entity.level, entity.getEyePosition(), Voicechat.SERVER_CONFIG.voiceChatDistance.get(), filter == null ? player -> true : player -> filter.test(new ServerPlayerImpl(player))), packet, null, null, null, SoundPacketEvent.SOURCE_PLUGIN);
+            server.broadcast(ServerWorldUtils.getPlayersInRange((ServerLevel) entity.level, entity.getEyePosition(), server.getBroadcastRange(), filter == null ? player -> true : player -> filter.test(new ServerPlayerImpl(player))), packet, null, null, null, SoundPacketEvent.SOURCE_PLUGIN);
         }
     }